How to make targeted offers to customers
The Self-Learning Response Model (SLRM) node enables you to build a model that you can continually update. Such updates are useful when building a model that assists with predicting which offers are most appropriate for customers and the probability of offers being accepted. These sorts of models are most beneficial in customer relationship management, such as marketing applications or call centers.
This example is based on a fictional banking company.
The marketing department wants to achieve more profitable results in future campaigns by matching the right offer of financial services to each customer.
In particular, the data science department identified the characteristics of customers who are most likely to respond favorably based on previous offers and responses and to promote the best current offer based on the results and now need to compute the best offering plan.